GenRaw v0.2.0 GenRaw

Types

data()

data() :: binary()

filter()

filter() :: function()

if_name()

if_name() :: binary()

opts()

opts() :: keyword()

pdu()

pdu() :: keyword(non_neg_integer() | binary())

status()

status() :: :open | :closed

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

close(pid)

close(pid()) :: :ok | {:error, any()}

Close a raw socket at GenRaw process pid.

interfaces()

interfaces() :: {:ok, [binary()]} | {:error, any()}

Retrieve a list of available interfaces.

Example

iex> GenRaw.interfaces
{:ok, ["lo", "eth0"]}

open(pid, opts \\ [])

open(pid(), opts()) :: :ok | {:ok, port()} | {:error, any()}

Open a raw socket with GenRaw process pid. This is equivalent to socket(AF_PACKET, SOCK_RAW, ETH_P_ALL). See man 2 socket for details.

To observe the resulting socket in a shell, try ss -elnp0.

Active mode

Setting active: true in opts will send data directly to the owner process. In this case, open/2 returns {:ok, port} on success. The data message format is

{port, {:data, data}}

where port is the port returned by open/2 and data is a binary.

receive(pid)

receive(pid()) :: {:ok, pdu()} | {:error, any()}

Receive a PDU. If no PDU is in queue, then {:error, :eagain} is returned.

Calling receive/1 on a GenRaw process opened with active: true will return {:error, :eopnotsupp}.

Example

iex> GenRaw.receive(pid)
{:error, :eagain}
iex> GenRaw.receive(pid)
{ :ok,
  <<255, 255, 255, 255, 255, 255,
    192, 255, 51, 192, 255, 51,
    0, 4,
    116,101,115,116
  >>
}

receive_parsed(pid, filter \\ fn _ -> true end)

receive_parsed(pid(), filter()) :: {:ok, pdu()} | {:error, any()}

Receive a parsed PDU.

When filter is provided, unmatched PDUs will be dropped until a matching PDU is found. When no PDU in the queue matches, {:error, :eagain} is returned.

Calling this function on a GenRaw process opened with active: true will return {:error, :eopnotsupp}.

Examples

iex> filter =
...>   &match?([dix: [_, {:src, <<0xc0ff33c0ff33::48>>}, _]], &1)
iex> 
iex> GenRaw.receive(pid, filter)
{ :ok, [
    dix: [
      dst: <<255, 255, 255, 255, 255, 255>>,
      src: <<192, 255, 51, 192, 255, 51>>,
      type: 4,
    ],
    data: "test",
  ]
}

send(pid, if_name, data)

send(pid(), if_name(), data()) :: :ok | {:error, any()}

Send data from interface if_name.

Sending a frame containing less than 14 bytes will return {:error, :emsgsize}.

On my system, sending more than 4193920 bytes will return {:error, :einval}, but I don't guard for it. Of course, this doesn't mean you should actually send 4193920 bytes, and interfaces are not normally configured to accept more than 1532 bytes anyway (MTU + 802.3 + 802.1Q + etc.).

Examples

iex> data  = "test"
iex> len   = byte_size(data)
iex> frame =
...>   <<0xffffffffffff::48,
...>     0xc0ff33c0ff33::48,
...>     test::bytes-size(len),
...>   >>
iex> GenRaw.send(pid, "eth0", frame)
:ok

start(opts \\ [])

start(opts()) :: {:ok, pid()} | {:error, any()}

Start a GenRaw process without linking it to the current process.

start_link(opts \\ [])

start_link(opts()) :: {:ok, pid()} | {:error, any()}

Start a GenRaw process, linking it to the current process.

status(pid)

status(pid()) :: status()

Get the status of GenRaw process pid.

Examples

iex> GenRaw.status(pid)
:closed
iex> GenRaw.open(pid)
:ok
iex> GenRaw.status(pid)
:open

stop(pid)

stop(pid()) :: :ok

Stop the GenRaw process at pid.